requirements.txt
# requirements.txt — Python Dependencies
#
# Install with: pip install -r requirements.txt
# (setup.sh does this automatically inside a venv)
#
# Tested with Python 3.10+

# Anthropic SDK — direct API access (optional if using the Claude Code proxy)
anthropic>=0.40.0

# OpenAI-compatible client — used to call the local Claude Code proxy
openai>=1.0.0

# Discord bot framework
discord.py>=2.3.0

# Environment variable loading
python-dotenv>=1.0.0

# YouTube audio/video (optional — for skills that reference YouTube)
yt-dlp>=2024.0.0

# HTTP requests (for Perplexity search and other API calls)
requests>=2.31.0

# Job scheduler (heartbeat loop)
schedule>=1.2.0
